Update to 16 - 11 - 24

+++ Ecommerce/EcommerceMiddleware/src/main/java/com/vegusa/middleware/repository/AuthTokenRepository.java
@@ -6,8 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AuthTokenRepository extends JpaRepository<AuthToken, Integer>
-{
+public interface AuthTokenRepository extends JpaRepository<AuthToken, Long> {
     @Query(value = "select * from AuthToken where IntegrationCompany = ?1",nativeQuery = true)
     AuthToken getAuthToken(String integrationCompany);
 }
